[no-issue] refactor: action 폴더 위치 변경
